third_party/blink/web_tests/external/wpt/css/css-page/parsing/size-invalid-expected.txt

This is a testharness.js-based test.
[FAIL] e.style['size'] = "initial" should not set the property value
  assert_equals: expected "" but got "initial"
[FAIL] e.style['size'] = "inherit" should not set the property value
  assert_equals: expected "" but got "inherit"
[FAIL] e.style['size'] = "revert" should not set the property value
  assert_equals: expected "" but got "revert"
[FAIL] e.style['size'] = "revert-layer" should not set the property value
  assert_equals: expected "" but got "revert-layer"
[FAIL] e.style['size'] = "unset" should not set the property value
  assert_equals: expected "" but got "unset"
[FAIL] e.style['size'] = "640px 480px" should not set the property value
  assert_equals: expected "" but got "640px 480px"
[FAIL] e.style['size'] = "8.5in 11in" should not set the property value
  assert_equals: expected "" but got "8.5in 11in"
[FAIL] e.style['size'] = "a4" should not set the property value
  assert_equals: expected "" but got "a4"
[FAIL] e.style['size'] = "3in 10in" should not set the property value
  assert_equals: expected "" but got "3in 10in"
[FAIL] e.style['size'] = "jis-b5" should not set the property value
  assert_equals: expected "" but got "jis-b5"
[FAIL] e.style['size'] = "auto" should not set the property value
  assert_equals: expected "" but got "auto"
[FAIL] e.style['size'] = "landscape" should not set the property value
  assert_equals: expected "" but got "landscape"
[FAIL] e.style['size'] = "letter" should not set the property value
  assert_equals: expected "" but got "letter"
[FAIL] e.style['size'] = "legal landscape" should not set the property value
  assert_equals: expected "" but got "legal landscape"
Harness: the test ran to completion.
